salsa20: internal refactoring
Unifies the `Salsa`/`SalsaCore` types, as one was previously just a thin
wrapper around the other.

This matches the structure of the `chacha20` crate.
