+++ salsa20/src/lib.rs
@@ -64,24 +64,13 @@ mod salsa;
 #[cfg(feature = "xsalsa20")]
 mod xsalsa;
 
+pub use crate::salsa::{Key, Nonce, Salsa, Salsa12, Salsa20, Salsa8};
+
 #[cfg(feature = "xsalsa20")]
-pub use self::xsalsa::{XNonce, XSalsa20};
+pub use crate::xsalsa::{XNonce, XSalsa20};
 
 #[cfg(feature = "hsalsa20")]
-pub use self::xsalsa::hsalsa20;
-
-use crate::{
-    block::Block,
-    rounds::{Rounds, R12, R20, R8},
-    salsa::SalsaCore,
-};
-use cipher::{
-    consts::{U32, U8},
-    stream::{
-        LoopError, NewStreamCipher, OverflowError, SeekNum, SyncStreamCipher, SyncStreamCipherSeek,
-    },
-};
-use core::convert::TryInto;
+pub use crate::xsalsa::hsalsa20;
 
 /// Size of a Salsa20 block in bytes
 pub const BLOCK_SIZE: usize = 64;
@@ -89,71 +78,8 @@ pub const BLOCK_SIZE: usize = 64;
 /// Size of a Salsa20 key in bytes
 pub const KEY_SIZE: usize = 32;
 
-/// Number of bytes in the Salsa20 IV
-const IV_SIZE: usize = 8;
-
 /// Number of 32-bit words in the Salsa20 state
 const STATE_WORDS: usize = 16;
 
 /// State initialization constant ("expand 32-byte k")
 const CONSTANTS: [u32; 4] = [0x6170_7865, 0x3320_646e, 0x7962_2d32, 0x6b20_6574];
-
-/// Salsa20/8 stream cipher
-/// (reduced-round variant of Salsa20 with 8 rounds, *not recommended*)
-pub type Salsa8 = Salsa<R8>;
-
-/// Salsa20/12 stream cipher
-/// (reduced-round variant of Salsa20 with 12 rounds, *not recommended*)
-pub type Salsa12 = Salsa<R12>;
-
-/// Salsa20/20 stream cipher
-/// (20 rounds; **recommended**)
-pub type Salsa20 = Salsa<R20>;
-
-/// Key type.
-///
-/// Implemented as an alias for [`GenericArray`].
-///
-/// (NOTE: all three round variants use the same key size)
-pub type Key = cipher::stream::Key<Salsa20>;
-
-/// Nonce type.
-///
-/// Implemented as an alias for [`GenericArray`].
-pub type Nonce = cipher::stream::Nonce<Salsa20>;
-
-/// The Salsa20 family of stream ciphers
-/// (implemented generically over a number of rounds).
-///
-/// We recommend you use the [`Salsa20`] (a.k.a. Salsa20/20) variant.
-#[derive(Debug)]
-pub struct Salsa<R: Rounds>(SalsaCore<R>);
-
-impl<R: Rounds> NewStreamCipher for Salsa<R> {
-    /// Key size in bytes
-    type KeySize = U32;
-
-    /// Nonce size in bytes
-    type NonceSize = U8;
-
-    fn new(key: &Key, nonce: &Nonce) -> Self {
-        let block = Block::new(key.as_slice().try_into().unwrap(), (*nonce).into());
-        Salsa(SalsaCore::new(block))
-    }
-}
-
-impl<R: Rounds> SyncStreamCipherSeek for Salsa<R> {
-    fn try_current_pos<T: SeekNum>(&self) -> Result<T, OverflowError> {
-        self.0.try_current_pos()
-    }
-
-    fn try_seek<T: SeekNum>(&mut self, pos: T) -> Result<(), LoopError> {
-        self.0.try_seek(pos)
-    }
-}
-
-impl<R: Rounds> SyncStreamCipher for Salsa<R> {
-    fn try_apply_keystream(&mut self, data: &mut [u8]) -> Result<(), LoopError> {
-        self.0.try_apply_keystream(data)
-    }
-}
